Refactor code to use composer and autoload. Add realization of patterns Strategy and Factory. Write comparison of them in file "Patterns comparison"
