+++ strategies/HeroCreationStrategy.php
@@ -1,15 +1,11 @@
 <?php
-//namespace models;
 
-use models\Fighter;
+namespace strategies;
+
 use models\Hero;
-use models\Killer;
-use models\Shooter;
-use models\Tank;
 use models\Weapon;
-use models\Wizard;
 
-class HeroCreator
+abstract class HeroCreationStrategy
 {
     private array $weapons = [];
 
@@ -43,21 +39,5 @@ class HeroCreator
         }
     }
 
-    public function createHero(string $class, string $name, float $health, float $power, float $stamina, ?Weapon $weapon): Hero
-    {
-        if ($health < 0 || $power < 0 || $stamina < 0) {
-            throw new InvalidArgumentException('Invalid numerical hero parameters');
-        }
-        if ($weapon === null) {
-            $weapon = new Weapon('Nothing', 0, 0);
-        }
-        return match ($class) {
-            'Shooter' => new Shooter($name, $health, $power, $stamina, $weapon),
-            'Fighter' => new Fighter($name, $health, $power, $stamina, $weapon),
-            'Killer' => new Killer($name, $health, $power, $stamina, $weapon),
-            'Wizard' => new Wizard($name, $health, $power, $stamina, $weapon),
-            'Tank' => new Tank($name, $health, $power, $stamina, $weapon),
-            default => throw new InvalidArgumentException('Invalid class')
-        };
-    }
+    public abstract function createHero(string $name, float $health, float $power, float $stamina, ?Weapon $weapon): Hero;
 }
